+++ src/main/java/io/quire/api/model/chat/ChatOwner.java
@@ -1,0 +1,8 @@
+package io.quire.api.model.chat;
+
+import io.quire.api.model.work.*;
+import io.swagger.annotations.*;
+
+@ApiModel()
+public class ChatOwner extends WorkOwner {
+}