+++ src/main/java/io/quire/api/model/chat/CreateChatBody.java
@@ -1,0 +1,11 @@
+package io.quire.api.model.chat;
+
+import io.quire.api.model.work.*;
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
+
+import java.util.List;
+
+@ApiModel()
+public class CreateChatBody extends CreateWorkBody {
+}