+++ src/main/java/io/quire/api/model/chat/UpdateChatBody.java
@@ -1,0 +1,9 @@
+package io.quire.api.model.chat;
+
+import io.quire.api.model.work.*;
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
+
+@ApiModel()
+public class UpdateChatBody extends UpdateWorkBody {
+}